refactor: Restructure exam questions module, update imports, and remove unused event and state files
